Drop leaderless parties
